+++ subsys/mgmt/mcumgr/lib/cmd/stat_mgmt/src/stat_mgmt.c
@@ -8,21 +8,94 @@
 #include <string.h>
 #include <stdio.h>
 
-#include "mgmt/mgmt.h"
+#include <stats/stats.h>
+#include <mgmt/mgmt.h>
+#include <stat_mgmt/stat_mgmt_config.h>
+#include <stat_mgmt/stat_mgmt.h>
 #include "cborattr/cborattr.h"
-#include "stat_mgmt/stat_mgmt.h"
-#include "stat_mgmt/stat_mgmt_impl.h"
-#include "stat_mgmt/stat_mgmt_config.h"
 
 static struct mgmt_handler stat_mgmt_handlers[];
 
-static int
-stat_mgmt_cb_encode(struct stat_mgmt_entry *entry, void *arg)
-{
-	CborEncoder *enc;
-	CborError err;
+typedef int stat_mgmt_foreach_entry_fn(CborEncoder *enc, struct stat_mgmt_entry *entry);
 
-	enc = arg;
+static int
+stat_mgmt_walk_cb(struct stats_hdr *hdr, void *arg, const char *name, uint16_t off);
+
+struct stat_mgmt_walk_arg {
+	stat_mgmt_foreach_entry_fn *cb;
+	CborEncoder *enc;
+};
+
+static int
+stat_mgmt_walk_cb(struct stats_hdr *hdr, void *arg, const char *name, uint16_t off)
+{
+	struct stat_mgmt_walk_arg *walk_arg;
+	struct stat_mgmt_entry entry;
+	void *stat_val;
+
+	walk_arg = arg;
+
+	stat_val = (uint8_t *)hdr + off;
+	switch (hdr->s_size) {
+	case sizeof(uint16_t):
+		entry.value = *(uint16_t *) stat_val;
+		break;
+	case sizeof(uint32_t):
+		entry.value = *(uint32_t *) stat_val;
+		break;
+	case sizeof(uint64_t):
+		entry.value = *(uint64_t *) stat_val;
+		break;
+	default:
+		return MGMT_ERR_EUNKNOWN;
+	}
+	entry.name = name;
+
+	return walk_arg->cb(walk_arg->enc, &entry);
+}
+
+static int
+stat_mgmt_get_group(int idx, const char **out_name)
+{
+	const struct stats_hdr *cur;
+	int i;
+
+	cur = NULL;
+	for (i = 0; i <= idx; i++) {
+		cur = stats_group_get_next(cur);
+		if (cur == NULL) {
+			return MGMT_ERR_ENOENT;
+		}
+	}
+
+	*out_name = cur->s_name;
+	return 0;
+}
+
+
+static int
+stat_mgmt_foreach_entry(const char *group_name, stat_mgmt_foreach_entry_fn *cb, CborEncoder *enc)
+{
+	struct stat_mgmt_walk_arg walk_arg;
+	struct stats_hdr *hdr;
+
+	hdr = stats_group_find(group_name);
+	if (hdr == NULL) {
+		return MGMT_ERR_ENOENT;
+	}
+
+	walk_arg = (struct stat_mgmt_walk_arg) {
+		.cb = cb,
+		.enc = enc
+	};
+
+	return stats_walk(hdr, stat_mgmt_walk_cb, &walk_arg);
+}
+
+static int
+stat_mgmt_cb_encode(CborEncoder *enc, struct stat_mgmt_entry *entry)
+{
+	CborError err;
 
 	err = 0;
 	err |= cbor_encode_text_stringz(enc, entry->name);
@@ -70,8 +143,7 @@ stat_mgmt_show(struct mgmt_ctxt *ctxt)
 	err |= cbor_encode_text_stringz(&ctxt->encoder, "fields");
 	err |= cbor_encoder_create_map(&ctxt->encoder, &map_enc, CborIndefiniteLength);
 
-	rc = stat_mgmt_impl_foreach_entry(stat_name, stat_mgmt_cb_encode,
-									  &map_enc);
+	rc = stat_mgmt_foreach_entry(stat_name, stat_mgmt_cb_encode, &map_enc);
 
 	err |= cbor_encoder_close_container(&ctxt->encoder, &map_enc);
 	if (err != 0) {
@@ -103,7 +175,7 @@ stat_mgmt_list(struct mgmt_ctxt *ctxt)
 	 * array.
 	 */
 	for (i = 0; ; i++) {
-		rc = stat_mgmt_impl_get_group(i, &group_name);
+		rc = stat_mgmt_get_group(i, &group_name);
 		if (rc == MGMT_ERR_ENOENT) {
 			/* No more stat groups. */
 			break;
